Add files via upload
this is using csv due to the original design of having dual camera setup side by side. Note this is the only method to having a dual camera setup. This method is faulty as it is limited to the raspberry internal gpu which causes the program to crash or be unresponsive. therefore, new software is being made to integrate a single camera stream which will need the use of a controller, in order to swap images. This will work in with the xbox controller which will have a periodic polling